extract `makeAbsoluteUrls` into utils/
