+++ packages/url-utils/lib/utils/index.js
@@ -4,6 +4,7 @@ module.exports = {
     deduplicateSubdirectory: require('./deduplicate-subdirectory'),
     getApiPath: require('./get-api-path'),
     getVersionPath: require('./get-version-path'),
+    htmlRelativeToAbsolute: require('./html-relative-to-absolute'),
     isSSL: require('./is-ssl'),
     relativeToAbsolute: require('./relative-to-absolute'),
     replacePermalink: require('./replace-permalink'),
